+++ Navigation/Navigation/Profile/ProfileViewController.swift
@@ -7,12 +7,14 @@
 
 import UIKit
 
+
 class ProfileViewController: UIViewController {
     
     lazy var tableView: UITableView = {
         let tableView = UITableView(frame: .zero, style: .grouped)
         tableView.translatesAutoresizingMaskIntoConstraints = false
         tableView.rowHeight = UITableView.automaticDimension
+        tableView.register(<#T##nib: UINib?##UINib?#>, forCellReuseIdentifier: <#T##String#>)
         
         return tableView
     }()
